+++ mirrord/cli/Cargo.toml
@@ -25,6 +25,7 @@ mirrord-config = { path = "../config" }
 clap.workspace = true
 tracing.workspace = true
 rand.workspace = true
+serde_json.workspace = true
 tracing-subscriber.workspace = true
 regex = "1.6.0"
 errno = "0.2"
@@ -34,6 +35,8 @@ reqwest.workspace = true
 semver = "1"
 const-random = "0.1.15"
 tokio.workspace = true
+kube.workspace = true
+k8s-openapi.workspace = true
 
 [target.'cfg(target_os = "macos")'.dependencies]
 mirrord-sip = { path = "../sip" }
