+++ mirrord/cli/src/config.rs
@@ -24,6 +24,10 @@ pub(super) enum Commands {
     /// Operator commands eg. setup
     #[command(hide = true)]
     Operator(Box<OperatorArgs>),
+
+    /// List targets/resources like pods/namespaces in json format    
+    #[command(hide = true, name = "ls")]
+    ListTargets(Box<ListTargetArgs>),
 }
 
 #[derive(Copy, Clone, PartialEq, Eq, PartialOrd, Ord, ValueEnum, Debug)]
@@ -216,4 +220,26 @@ pub(super) enum OperatorCommand {
         #[arg(short, long, default_value = "mirrord")]
         namespace: OperatorNamespace,
     },
+}
+
+#[derive(ValueEnum, Clone, Debug)]
+pub enum Format {
+    Json,
+}
+
+#[derive(Args, Debug)]
+pub(super) struct ListTargetArgs {
+    /// Specify the format of the output.
+    #[arg(
+        short = 'o',
+        long = "output",
+        value_name = "FORMAT",
+        value_enum,
+        default_value_t = Format::Json
+    )]
+    pub output: Format,
+
+    /// Specify the namespace to list targets in.
+    #[arg(short = 'n', long = "namespace")]
+    pub namespace: Option<String>,
 }
