+++ mirrord/layer/Cargo.toml
@@ -43,7 +43,7 @@ thiserror.workspace = true
 trust-dns-resolver.workspace = true
 rand = "0.8"
 fancy-regex = "0.10"
-regex = "1"
+regex.workspace = true
 errno = "0.2"
 async-trait.workspace = true
 socket2 = "0.4"
